+++ R/emmGrid-methods.R
@@ -910,9 +910,10 @@ emm_defaults = list (
 #' degrees-of-freedom method is inherited from the previous method at
 #' the time of re-gridding.
 #'
-#' @note Another way to use \code{regrid} is to supply a \code{transform} 
+#' @note Another way to use \code{regrid} is to supply a \code{regrid} 
 #'   argument to \code{\link{ref_grid}} (either directly of indirectly via
-#'   \code{\link{emmeans}}). This is often a simpler approach if the reference
+#'   \code{\link{emmeans}}), in which case its value is passed to \code{regrid} as
+#'   \code{transform}. This is often a simpler approach if the reference
 #'   grid has not already been constructed.
 #'
 #' @return An \code{emmGrid} object with the requested changes
@@ -930,7 +931,7 @@ emm_defaults = list (
 #' (emm2 <- emmeans(regrid(rg, transform = "response"), "source"))
 #' pairs(emm2)  ## We obtain DIFFERENCES
 #' # Same result, useful if we hadn't already created 'rg'
-#' # emm2 <- emmeans(pigs.lm, "source", transform = "response")
+#' # emm2 <- emmeans(pigs.lm, "source", regrid = "response")
 #' 
 #' # Simulate a sample of regression coefficients
 #' set.seed(2.71828)